train_yolo.py
"""
Simplified training script using ultralytics YOLOv9
with custom EMA-SlideLoss integration
"""
from ultralytics import YOLO
import os


def train_ddeyolov9():
    """
    Train DDEYOLOv9 using ultralytics framework
    
    Note: To fully integrate DRNELAN4 and DCNv4-Dyhead, you would need to:
    1. Modify YOLOv9's architecture file to replace RepNCSPELAN4 with DRNELAN4
    2. Replace the detection head with DCNv4-Dyhead
    3. Integrate EMA-SlideLoss into the loss calculation
    
    For now, this script trains YOLOv9 with the dataset and shows the structure.
    """
    
    # Paths
    data_yaml = 'data.yaml'
    model_path = 'yolov9e.pt'  # YOLOv9-E model
    
    # Create data.yaml file
    data_config = {
        'path': os.path.abspath('.'),
        'train': 'images',
        'val': 'images',  # In practice, split your data
        'test': 'images',
        'nc': 5,  # Number of classes
        'names': ['Normal-Fish', 'PH abnormal-Fish', 'Low temperature-Fish', 
                 'High temperature-Fish', 'Hypoxia-Fish']
    }
    
    import yaml
    with open(data_yaml, 'w') as f:
        yaml.dump(data_config, f)
    
    # Load model
    model = YOLO(model_path)
    
    # Train
    results = model.train(
        data=data_yaml,
        epochs=200,
        imgsz=640,
        batch=8,
        lr0=0.01,
        device=0 if os.path.exists('/dev/cuda') else 'cpu',
        project='runs/train',
        name='ddeyolov9',
        exist_ok=True,
        pretrained=True,
        optimizer='SGD',
        verbose=True,
        seed=42,
        deterministic=True,
        single_cls=False,
        rect=False,
        cos_lr=True,
        close_mosaic=10,
        resume=False,
        amp=True,
        fraction=1.0,
        profile=False,
        freeze=None,
        # Custom parameters
        # To integrate EMA-SlideLoss, you would modify the loss function in YOLO's code
    )
    
    # Validate
    metrics = model.val()
    print(f"\nValidation Results:")
    print(f"mAP50: {metrics.box.map50:.4f}")
    print(f"mAP50-95: {metrics.box.map:.4f}")
    print(f"Precision: {metrics.box.mp:.4f}")
    print(f"Recall: {metrics.box.mr:.4f}")
    
    # Export model
    model.export(format='onnx')
    print("\nModel exported to ONNX format")
    
    return results


if __name__ == '__main__':
    train_ddeyolov9()

