requirements.txt
torch>=1.10.0
torchvision>=0.11.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy<2.0
Pillow>=8.3.0
albumentations>=1.1.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.7.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

